Feature complete, ready to submit for review

+++ app_catacity.py
@@ -13,9 +13,9 @@ import random, string
 from oauth2client.client import flow_from_clientsecrets
 from oauth2client.client import FlowExchangeError
 import httplib2
+import requests  # improved over urllib2
 import json
 from flask import make_response
-import requests  # improved over urllib2
 
 #auth = HTTPBasicAuth()
 CLIENT_ID = json.loads(
@@ -27,16 +27,36 @@ Base.metadata.bind = engine
 DBSession = sessionmaker(bind = engine)
 session = DBSession()
 
+#       #######  #####    ###   #     #
+#       #     # #     #    #    ##    #
+#       #     # #          #    # #   #
+#       #     # #  ####    #    #  #  #
+#       #     # #     #    #    #   # #
+#       #     # #     #    #    #    ##
+####### #######  #####    ###   #     #
+
 #------------------------------------------------------
 #  Login
 #------------------------------------------------------
 @app.route('/login')
 def showLogin():
-	# Create anti-forgery state token
-	state = ''.join(random.choice(
-		string.ascii_uppercase + string.digits) for x in xrange(32))
-	login_session['state'] = state
-	return render_template('login.html', STATE=state)
+	if 'username' not in login_session:
+		# Create anti-forgery state token
+		state = ''.join(random.choice(
+			string.ascii_uppercase + string.digits) for x in xrange(32))
+		login_session['state'] = state
+		return render_template('login.html', STATE=state)
+	else:
+		return render_template('logout.html')
+
+
+ #####  ####### #######  #####  #       #######
+#     # #     # #     # #     # #       #
+#       #     # #     # #       #       #
+#  #### #     # #     # #  #### #       #####
+#     # #     # #     # #     # #       #
+#     # #     # #     # #     # #       #
+ #####  ####### #######  #####  ####### #######
 
 #------------------------------------------------------
 #  Google OAuth2
@@ -51,18 +71,14 @@ def gconnect():
  
   code = request.data  # This is the one-time use code from Google via client
 
-#  return("<p>Got this far in gconnect.</p>State: " + login_session['state'])
-
   # Upgrade the authorization code into a credentials object
   try:
     oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
     oauth_flow.redirect_uri = 'postmessage'
     credentials = oauth_flow.step2_exchange(code) # exchanges auth code for credentials object
-#    return("<p>Succeeded getting credentials</p>")
   except FlowExchangeError:
     response = make_response(json.dumps('Failed to upgrade the auth code.'), 401)
     response.headers['Content-Type'] = 'application/json'
-#    return("<p>Failed getting credentials</p>")
     return response
 
   # Verify access token is valid by sending it to Google for a check
@@ -74,7 +90,7 @@ def gconnect():
 
   # If there was an error in the access token info, abort.
   if result.get('error') is not None:
-    response = make_response(json.dumps(result.get('error')), 500)
+    response = make_response(json.dumps(result.get('error')), 500) 
     response.headers['Content-Type'] = 'application/json'
     return response
 
@@ -82,7 +98,7 @@ def gconnect():
   gplus_id = credentials.id_token['sub']
   if result['user_id'] != gplus_id:
     response = make_response(json.dumps(
-      "Token UserID does not match given UserID"), 401)
+      "Token UserID does not match given UserID"), 401) 
     response.headers['Content-Type'] = 'application/json'
     return response
 
@@ -103,12 +119,6 @@ def gconnect():
 
   # Passed all checks!
 
-  # Store the access token in the session for later use.
-  #login_session['credentials'] = credentials
-  login_session['provider'] = 'google'
-  login_session['access_token'] = credentials.access_token
-  login_session['gplus_id'] = gplus_id
-
   # Get User info
   userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
   params = {'access_token': credentials.access_token, 'alt':'json'}
@@ -116,15 +126,70 @@ def gconnect():
   data = json.loads(answer.text)
   # "data" now contains all info listed in Google's getOpenIdConnect response
 
+  # Store the access token in the session for later use.
+  login_session['provider'] = 'google'
+  login_session['access_token'] = credentials.access_token
+  login_session['gplus_id'] = gplus_id
   login_session['username'] = data["name"]
   login_session['picture'] = data["picture"]
   login_session['email'] = data["email"]
 
+  # see if user exists, if not then create
+  user_id = getUserID(login_session['email'])
+  if not user_id:
+  	user_id = createUser(login_session)
+  login_session['user_id'] = user_id
+
   test =  '<p>Username: ' + login_session['username'] + '</p>'
   test += '<p>Email: '    + login_session['email'] + '</p>'
   test += '<p>Picture: '  + login_session['picture'] + '</p>'
-  flash("You are no logged in as %s" %login_session['username'])
+  test += '<p>User ID: '  + str(login_session['user_id']) + '</p>'
+  flash("You are now logged in as %s" %login_session['username'])
   return test
+
+
+
+                ##############
+                ##          ##
+                ##          ##
+                ##          ##
+                ##          ##
+            ######          ######
+              ##              ##
+                ##          ##
+                  ##      ##
+                    ##  ##
+                      ##
+
+
+# DISCONNECT - Revoke a current user's token and reset their login_session
+@app.route("/gdisconnect")
+def gdisconnect():
+  # Only disconnect a connected user.
+  access_token = login_session.get('access_token')
+  if access_token is None:
+    response = make_response(json.dumps('Current user not connected.'), 401)
+    response.headers['Content-Type'] = 'application/json'
+    return response
+
+  # Execute HTTP GET request to revoke current token
+  r = requests.post('https://accounts.google.com/o/oauth2/revoke',
+    params={'token': access_token},
+    headers = {'content-type': 'application/x-www-form-urlencoded'})
+  # From Google:
+    # Note: Following a successful revocation response,
+    # it might take some time before the revocation has full effect.
+
+  del login_session['access_token']
+  del login_session['gplus_id']
+  del login_session['username']
+  del login_session['email']
+  del login_session['picture']
+  #response = make_response(json.dumps('Successfully disconnected.'), 200)
+  #response.headers['Content-Type'] = 'application/json'
+  flash("Successfully logged out.")
+  return redirect("/", code=302)
+
 
 ########################################
 # Debug pages
@@ -144,6 +209,14 @@ def showState():
 	login_session['state'] = state
 	return "Session state: %s" %login_session['state']
 ########################################
+
+#     #   ###   ####### #     #
+#     #    #    #       #  #  #
+#     #    #    #       #  #  #
+#     #    #    #####   #  #  #
+ #   #     #    #       #  #  #
+  # #      #    #       #  #  #
+   #      ###   #######  ## ##
 
 #------------------------------------------------------
 #  LIST Categories
@@ -171,7 +244,31 @@ def showItems(category_id):
 def viewItem(category_id, item_id):
 	category = session.query(Category).filter_by(id = category_id).one()
 	item = session.query(Item).filter_by(id = item_id, category_id = category_id).one()
-	return render_template('viewItem.html', category = category, item = item)
+	if 'username' not in login_session or item.acct_id != login_session['user_id']:
+		return render_template('publicViewItem.html', category = category, item = item)
+	#if 'username' not in login_session:# or item.acct_id != login_session['user_id']:
+	#	return render_template('publicViewItem.html', category = category, item = item)
+	#elif item.acct_id != login_session['user_id']:
+	#	return render_template('publicViewItem.html', category = category, item = item)
+	else:
+		return render_template('viewItem.html', category = category, item = item)
+
+#------------------------------------------------------
+#  VIEW All of My Items
+#------------------------------------------------------
+@app.route('/myitems/')
+def showMyItems():
+	items = session.query(Item).filter_by(acct_id = login_session['user_id']).all()
+	return render_template('showMyItems.html', items = items)
+
+
+####### ######    ###   #######
+#       #     #    #       #
+#       #     #    #       #
+#####   #     #    #       #
+#       #     #    #       #
+#       #     #    #       #
+####### ######    ###      #
 
 #------------------------------------------------------
 #  NEW Item
@@ -179,11 +276,13 @@ def viewItem(category_id, item_id):
 @app.route('/category/<int:category_id>/new/', methods=['GET', 'POST'])
 def newItem(category_id):
 #	return "Test string for: /category/<int:category_id>/new/"
+	if 'username' not in login_session:
+		return redirect('/login')
 	if request.method == 'POST':
 		newItem = Item(title = request.form['title'],
 			description = request.form['description'],
 			category_id = category_id,
-			acct_id = 1) 
+			acct_id = login_session['user_id']) 
 #TODO: replace account_id value with the authenticated user_id		
 		session.add(newItem)
 		session.commit()
@@ -200,8 +299,16 @@ def newItem(category_id):
 @app.route('/category/<int:category_id>/item/<int:item_id>/edit/', methods=['GET', 'POST'])
 def editItem(category_id, item_id):
 #	return "Test string for: /category/<int:category_id>/item/<int:item_id>/edit/"
-
 	selItem = session.query(Item).filter_by(id = item_id, category_id = category_id).one()
+
+	#if selItem.id != login_session['user_id']:
+	#	test += '<p>Item: '  + login_session['picture'] + '</p>'
+	#	test += '<p>User ID: '  + str(login_session['user_id']) + '</p>'
+  	#	return test
+
+	if 'username' not in login_session or selItem.acct_id != login_session['user_id']:
+		return redirect('/login')
+
 	if request.method == 'POST':
 		if selItem != [] :
 			selItem.title = request.form['title']
@@ -223,6 +330,10 @@ def editItem(category_id, item_id):
 @app.route('/category/<int:category_id>/item/<int:item_id>/delete/', methods=['GET', 'POST'])
 def deleteItem(category_id, item_id):
 	deleteItem = session.query(Item).filter_by(id = item_id, category_id = category_id).one()
+
+	if 'username' not in login_session or deleteItem.acct_id != login_session['user_id']:
+		return redirect('/login')
+
 	if request.method == 'POST':
 		if deleteItem != [] :
 			session.delete(deleteItem)
@@ -233,6 +344,45 @@ def deleteItem(category_id, item_id):
 		category = session.query(Category).filter_by(id = category_id).one()
 		return render_template('deleteItem.html', category = category, item = deleteItem)
 
+
+######  ####### ######  #     #   ###    #####   #####    ###   ####### #     #
+#     # #       #     # ##   ##    #    #     # #     #    #    #     # ##    #
+#     # #       #     # # # # #    #    #       #          #    #     # # #   #
+######  #####   ######  #  #  #    #     #####   #####     #    #     # #  #  #
+#       #       #   #   #     #    #          #       #    #    #     # #   # #
+#       #       #    #  #     #    #    #     # #     #    #    #     # #    ##
+#       ####### #     # #     #   ###    #####   #####    ###   ####### #     #
+
+#------------------------------------------------------
+#  Permissions System
+#------------------------------------------------------
+def createUser(login_session):
+	newUser = User(name = login_session['username'],
+		email = login_session['email'])
+	session.add(newUser)
+	session.commit()
+	user = session.query(User).filter_by(email = login_session['email']).one()
+	return user.id
+
+def getUserID(email):
+	try:
+		user = session.query(User).filter_by(email = email).one()
+		return user.id
+	except:
+		return None
+
+def getUserInfo(user_id):
+	user = session.query(User).filter_by(id = user_id).one()
+	return user
+
+
+      #  #####  ####### #     #
+      # #     # #     # ##    #
+      # #       #     # # #   #
+      #  #####  #     # #  #  #
+#     #       # #     # #   # #
+#     # #     # #     # #    ##
+ #####   #####  ####### #     #
 
 #------------------------------------------------------
 #  JSON Endpoint
